Toggle and like managers.
