+++ app/src/main/java/com/alexbezhan/instagram/activities/profile/ProfileViewModel.kt
@@ -3,16 +3,15 @@ package com.alexbezhan.instagram.activities.profile
 import android.arch.lifecycle.LiveData
 import android.arch.lifecycle.Transformations
 import com.alexbezhan.instagram.activities.BaseViewModel
-import com.alexbezhan.instagram.activities.DefaultFollowListener
-import com.alexbezhan.instagram.activities.FollowListener
+import com.alexbezhan.instagram.activities.FollowManager
 import com.alexbezhan.instagram.activities.asUser
 import com.alexbezhan.instagram.models.User
 import com.alexbezhan.instagram.utils.firebase.FirebaseHelper.currentUid
 import com.alexbezhan.instagram.utils.firebase.FirebaseHelper.database
 import com.alexbezhan.instagram.utils.livedata.FirebaseLiveData
 
-class ProfileViewModel : BaseViewModel(), FollowListener {
-    private val followListener = DefaultFollowListener(setErrorOnFailureListener)
+class ProfileViewModel : BaseViewModel() {
+    private val followListener = FollowManager()
 
     var anotherUser: LiveData<User>? = null
 
@@ -30,6 +29,6 @@ class ProfileViewModel : BaseViewModel(), FollowListener {
         it.children.map { it.getValue(String::class.java)!! }
     }
 
-    override fun toggleFollow(currentUser: User, uid: String) =
-        followListener.toggleFollow(currentUser, uid)
+    fun toggleFollow(currentUser: User, uid: String) =
+        followListener.toggleFollow(currentUser, uid, setErrorOnFailureListener)
 }
