+++ app/src/main/java/com/alexbezhan/instagram/activities/profile/friends/AddFriendsViewModel.kt
@@ -3,15 +3,14 @@ package com.alexbezhan.instagram.activities.profile.friends
 import android.arch.lifecycle.LiveData
 import android.arch.lifecycle.Transformations
 import com.alexbezhan.instagram.activities.BaseViewModel
-import com.alexbezhan.instagram.activities.DefaultFollowListener
-import com.alexbezhan.instagram.activities.FollowListener
+import com.alexbezhan.instagram.activities.FollowManager
 import com.alexbezhan.instagram.activities.asUser
 import com.alexbezhan.instagram.models.User
 import com.alexbezhan.instagram.utils.firebase.FirebaseHelper
 import com.alexbezhan.instagram.utils.livedata.FirebaseLiveData
 
-class AddFriendsViewModel : BaseViewModel(), FollowListener {
-    private val followListener = DefaultFollowListener(setErrorOnFailureListener)
+class AddFriendsViewModel : BaseViewModel() {
+    private val followManager = FollowManager()
 
     val userAndFriends: LiveData<Pair<User, List<User>>> = Transformations.map(
             FirebaseLiveData(FirebaseHelper.database.child("users"))) {
@@ -21,7 +20,7 @@ class AddFriendsViewModel : BaseViewModel(), FollowListener {
         userList.first() to otherUsersList
     }
 
-    override fun toggleFollow(currentUser: User, uid: String) =
-            followListener.toggleFollow(currentUser, uid)
+    fun toggleFollow(currentUser: User, uid: String) =
+            followManager.toggleFollow(currentUser, uid, setErrorOnFailureListener)
 
 }
